1。将规格改成克重和宽度分开

+++ Source/LJH.Inventory.BusinessModel/SpecificationHelper.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using LJH.GeneralLibrary;
 
 namespace LJH.Inventory.BusinessModel
 {
@@ -10,7 +11,7 @@ namespace LJH.Inventory.BusinessModel
         /// <summary>
         /// 从产品规格中获取标定的厚度
         /// </summary>
-        public static decimal? GetWritten克重(string specification)
+        public static int? GetWritten克重(string specification)
         {
             if (!string.IsNullOrEmpty(specification))
             {
@@ -18,7 +19,7 @@ namespace LJH.Inventory.BusinessModel
                 if (temp.Length >= 1)
                 {
                     decimal ret = 0;
-                    if (decimal.TryParse(temp[0], out ret) && ret > 0) return ret;
+                    if (decimal.TryParse(temp[0], out ret) && ret > 0) return (int)ret;
                 }
             }
             return null;
@@ -27,7 +28,7 @@ namespace LJH.Inventory.BusinessModel
         /// <summary>
         /// 从产品规格中获取标定的宽度
         /// </summary>
-        public static decimal? GetWrittenWidth(string specification)
+        public static int? GetWrittenWidth(string specification)
         {
             if (!string.IsNullOrEmpty(specification))
             {
@@ -35,7 +36,7 @@ namespace LJH.Inventory.BusinessModel
                 if (temp.Length >= 2)
                 {
                     decimal ret = 0;
-                    if (decimal.TryParse(temp[1], out ret) && ret > 0) return ret;
+                    if (decimal.TryParse(temp[1], out ret) && ret > 0) return (int)ret;
                 }
             }
             return null;
